feat(auth): enhance authentication flow with session management and error handling

+++ Rasyonet_Intern.API/Settings/JwtSettings.cs
@@ -8,6 +8,6 @@ namespace Rasyonet_Intern.API.Settings
         public string Issuer { get; set; } = String.Empty;
         public string Audience { get; set; } = String.Empty;
         public string SecretKey { get; set; } = String.Empty;
-        public int ExpirationInMinutes { get; set; }
+        public double ExpirationInMinutes { get; set; }
     }
 }
